+++ client/src/components/ContentNavigator/const.ts
@@ -28,7 +28,8 @@ export const ROOT_FOLDER = {
 };
 
 export const FILE_TYPE = "file";
-export const FILE_TYPES = [FILE_TYPE];
+export const DATAFLOW_TYPE = "dataFlow";
+export const FILE_TYPES = [FILE_TYPE, DATAFLOW_TYPE];
 export const FOLDER_TYPE = "folder";
 export const MYFOLDER_TYPE = "myFolder";
 export const TRASH_FOLDER_TYPE = "trashFolder";
@@ -61,6 +62,7 @@ export const Messages = {
   FileDropError: l10n.t('Unable to drop item "{name}".'),
   FileOpenError: l10n.t("The file type is unsupported."),
   FileRestoreError: l10n.t("Unable to restore file."),
+  FileUploadError: l10n.t("Unable to upload files."),
   FileValidationError: l10n.t("Invalid file name."),
   FolderDeletionError: l10n.t("Unable to delete folder."),
   FolderRestoreError: l10n.t("Unable to restore folder."),
@@ -85,7 +87,7 @@ export const Messages = {
   ),
   ConvertNotebookToFlowPrompt: l10n.t("Enter a name for the new .flw file"),
   NotebookToFlowConversionSuccess: l10n.t(
-    "The notebook has been successfully converted to a flow. You can now open it in SAS Studio.",
+    "The notebook has been successfully converted to a flow and saved into the following folder: {folderName}. You can now open it in SAS Studio.",
   ),
   NotebookToFlowConversionError: l10n.t(
     "Error converting the notebook file to .flw format.",