+++ client/src/components/ContentNavigator/utils.ts
@@ -1,9 +1,9 @@
 // Copyright © 2023, SAS Institute Inc., Cary, NC, USA.  All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { Uri } from "vscode";
+import { SnippetString, Uri } from "vscode";
 
 import {
-  FILE_TYPE,
+  FILE_TYPES,
   FOLDER_TYPE,
   FOLDER_TYPES,
   TRASH_FOLDER_TYPE,
@@ -70,7 +70,7 @@ export const resourceType = (item: ContentItem): string | undefined => {
     actions.push("removeFromFavorites");
   } else if (
     item.type !== "reference" &&
-    [FOLDER_TYPE, FILE_TYPE].includes(type) &&
+    [FOLDER_TYPE, ...FILE_TYPES].includes(type) &&
     !isRecycled
   ) {
     actions.push("addToFavorites");
@@ -79,6 +79,10 @@ export const resourceType = (item: ContentItem): string | undefined => {
   // if item is a notebook file add action
   if (item?.name?.endsWith(".sasnb")) {
     actions.push("convertNotebookToFlow");
+  }
+
+  if (!isContainer(item)) {
+    actions.push("allowDownload");
   }
 
   if (actions.length === 0) {
@@ -111,3 +115,27 @@ export const isItemInRecycleBin = (item: ContentItem): boolean =>
   !!item && item.flags?.isInRecycleBin;
 
 export const isContentItem = (item): item is ContentItem => isValidItem(item);
+
+// A document uses uppercase letters _if_ are no words
+// (where word means gte 3 characters) that are lowercase.
+const documentUsesUppercase = (documentContent: string) =>
+  documentContent &&
+  !documentContent
+    // Exclude anything in quotes from our calculations
+    .replace(/('|")([^('|")]*)('|")/g, "")
+    .match(/([a-z]{3,})\S/g);
+
+export const getFileStatement = (
+  contentItemName: string,
+  documentContent: string,
+  fileFolderPath: string,
+): SnippetString => {
+  const usesUppercase = documentUsesUppercase(documentContent);
+  const cmd = "filename ${1:fileref} filesrvc folderpath='$1' filename='$2';\n";
+
+  return new SnippetString(
+    (usesUppercase ? cmd.toUpperCase() : cmd)
+      .replace("$1", fileFolderPath.replace(/'/g, "''"))
+      .replace("$2", contentItemName.replace(/'/g, "''")),
+  );
+};