+++ client/src/components/ResultPanel/ResultPanel.ts
@@ -2,14 +2,28 @@
 // SPDX-License-Identifier: Apache-2.0
 import { Uri, ViewColumn, WebviewPanel, l10n, window } from "vscode";
 
-import {
-  isSideResultEnabled,
-  isSinglePanelEnabled,
-} from "./Helper/SettingHelper";
+import { v4 } from "uuid";
+
+import { isSideResultEnabled, isSinglePanelEnabled } from "../utils/settings";
 
 let resultPanel: WebviewPanel | undefined;
+export const resultPanels: Record<string, WebviewPanel> = {};
 
 export const showResult = (html: string, uri?: Uri, title?: string) => {
+  html = html
+    // Inject vscode context into our results html body
+    .replace(
+      "<body ",
+      `<body data-vscode-context='${JSON.stringify({
+        preventDefaultContextMenuItems: true,
+      })}' `,
+    )
+    // Make sure the html and body take up the full height of the parent
+    // iframe so that the context menu is clickable anywhere on the page
+    .replace(
+      "</head>",
+      "<style>html,body { height: 100% !important; }</style></head>",
+    );
   const sideResult = isSideResultEnabled();
   const singlePanel = isSinglePanelEnabled();
   if (!title) {
@@ -17,8 +31,9 @@ export const showResult = (html: string, uri?: Uri, title?: string) => {
   }
 
   if (!singlePanel || !resultPanel) {
+    const resultPanelId = `SASResultPanel-${v4()}`;
     resultPanel = window.createWebviewPanel(
-      "SASSession", // Identifies the type of the webview. Used internally
+      resultPanelId, // Identifies the type of the webview. Used internally
       title, // Title of the panel displayed to the user
       {
         preserveFocus: true,
@@ -26,7 +41,13 @@ export const showResult = (html: string, uri?: Uri, title?: string) => {
       }, // Editor column to show the new webview panel in.
       {}, // Webview options. More on these later.
     );
-    resultPanel.onDidDispose(() => (resultPanel = undefined));
+    resultPanel.onDidDispose(
+      ((id) => () => {
+        delete resultPanels[id];
+        resultPanel = undefined;
+      })(resultPanelId),
+    );
+    resultPanels[resultPanelId] = resultPanel;
   } else {
     const editor = uri
       ? window.visibleTextEditors.find(